Move chrono interop utils to new library
This structure makes more sense and makes the library easier to
understand.
