+++ au/au_test.cc
@@ -31,30 +31,7 @@
 
 using ::testing::StaticAssertTypeEq;
 
-using namespace std::chrono_literals;
-
 namespace au {
-
-TEST(DurationQuantity, InterconvertsWithExactlyEquivalentChronoDuration) {
-    constexpr QuantityD<Seconds> from_chrono = std::chrono::duration<double>{1.23};
-    EXPECT_THAT(from_chrono, SameTypeAndValue(seconds(1.23)));
-
-    constexpr auto val = std::chrono::nanoseconds::rep{456};
-    constexpr std::chrono::nanoseconds from_au = nano(seconds)(val);
-    EXPECT_THAT(from_au.count(), SameTypeAndValue(val));
-}
-
-TEST(DurationQuantity, InterconvertsWithIndirectlyEquivalentChronoDuration) {
-    constexpr QuantityD<Seconds> from_chrono = as_quantity(1234ms);
-    EXPECT_THAT(from_chrono, SameTypeAndValue(seconds(1.234)));
-}
-
-TEST(AsChronoDuration, ProducesExpectedResults) {
-    constexpr auto original = milli(seconds)(12.3f);
-    constexpr auto result = as_chrono_duration(original);
-    EXPECT_THAT(result.count(), SameTypeAndValue(12.3f));
-    EXPECT_THAT(as_quantity(result), QuantityEquivalent(original));
-}
 
 TEST(Conversions, SupportIntMHzToU32Hz) {
     constexpr QuantityU32<Hertz> freq = mega(hertz)(40);
